Cleaned up the reporting controllers

+++ ripple-core/src/main/java/org/rippleosi/patient/summary/search/NotConfiguredPatientSearch.java
@@ -74,7 +74,7 @@ public class NotConfiguredPatientSearch implements PatientSearch {
     }
 
     @Override
-    public Integer findPatientCountByDepartment(String department) {
+    public Long findPatientCountByDepartment(String department) {
         throw ConfigurationException.unimplementedTransaction(PatientSearch.class);
     }
 }
